+++ lib/src/xml/utils/prefix_name.dart
@@ -1,6 +1,7 @@
 library xml.utils.prefix_name;
 
 import 'name.dart';
+import 'namespace.dart';
 
 /// An XML entity name with a prefix.
 class XmlPrefixName extends XmlName {
@@ -14,16 +15,7 @@ class XmlPrefixName extends XmlName {
   final String qualified;
 
   @override
-  String get namespaceUri {
-    for (var node = parent; node != null; node = node.parent) {
-      for (final attribute in node.attributes) {
-        if (attribute.name.prefix == xmlns && attribute.name.local == prefix) {
-          return attribute.value;
-        }
-      }
-    }
-    return null;
-  }
+  String get namespaceUri => lookupAttribute(parent, xmlns, prefix)?.value;
 
   XmlPrefixName(this.prefix, this.local, this.qualified) : super.internal();
 }